Migrate homemanager from standalone to nix module
zsh can not be set as default shell in standalone home manager. There
are some workarounds such as [custom scripts for chsh](https://discourse.nixos.org/t/unable-to-set-or-use-zsh-shell-from-standalone-home-manager-using-flakes/29269/8). In
addition it appears that hyprland and other goodies that require 
system level settings are also not possible.

Test nixmod with zsh

Switch bash to zsh and make zsh default shell.