Expand zsh config
oh-my-zsh does not seem useful. There is an alternative in nix to
directly use [plugins](https://nixos.wiki/wiki/Zsh). Powerlevel10k
is added as plugin as shown in the example. The p10k.zsh document
is to be added to nix-config and several default plugins are
disabled. A symlink is needed to ~/.p10k.zsh. Should be fixed later.

powerlevel10k requires meslo-lg-nfs font. It is added as a system
setting and should likely not be required as a system setting.

zsh-vim-mode adds vim keybindings and surround functionality.