vpn
Windscribe does not appear to have a nixpkg. However it can be worked
around with [wireguard](https://discourse.nixos.org/t/how-to-install-windscribe-vpn/39007/4).

Setting up a user system was too much trouble with permissions. It
appears that certain system services are required and wireguard wants
sudo. Adding wg-quick and some others to sudoers list with a custom
vpn user and group did not help.

Currently a sleep is needed because it depends on other services
that are not yet loaded at that time. A sleep of a second should
be enough but to be safe 2 seconds is used.